feat(ng-add): Improved comments and readability of code [squash-Me]
