--- apps/demos/src/examples.module.ts
@@ -1,140 +1,0 @@
-/**
- * @license
- * Copyright 2019 Dynatrace LLC
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// tslint:disable: max-file-line-count
-
-import { NgModule } from '@angular/core';
-import {
-  DtAlertExamplesModule,
-  DtAutocompleteExamplesModule,
-  DtBarIndicatorExamplesModule,
-  DtBreadcrumbsExamplesModule,
-  DtButtonExamplesModule,
-  DtButtonGroupExamplesModule,
-  DtCardExamplesModule,
-  DtChartExamplesModule,
-  DtCheckboxExamplesModule,
-  DtConfirmationDialogExamplesModule,
-  DtConsumptionExamplesModule,
-  DtContainerBreakpointObserverExamplesModule,
-  DtContextDialogExamplesModule,
-  DtCopyToClipboardExamplesModule,
-  DtCtaCardExamplesModule,
-  DtDrawerExamplesModule,
-  DtEmptyStateExamplesModule,
-  DtEventChartExamplesModule,
-  DtExpandablePanelExamplesModule,
-  DtExpandableSectionExamplesModule,
-  DtExpandableTextExamplesModule,
-  DtFilterFieldExamplesModule,
-  DtFormFieldExamplesModule,
-  DtFormattersExamplesModule,
-  DtHighlightExamplesModule,
-  DtIconExamplesModule,
-  DtInfoGroupExamplesModule,
-  DtInlineEditorExamplesModule,
-  DtInputExamplesModule,
-  DtKeyValueListExamplesModule,
-  DtLegendExamplesModule,
-  DtLinkExamplesModule,
-  DtLoadingDistractorExamplesModule,
-  DtMenuExamplesModule,
-  DtMicroChartExamplesModule,
-  DtOverlayExamplesModule,
-  DtPaginationExamplesModule,
-  DtProgressBarExamplesModule,
-  DtProgressCircleExamplesModule,
-  DtRadioExamplesModule,
-  DtExamplesSecondaryNAvModule,
-  DtExamplesSelectModule,
-  DtExamplesShowMoreModule,
-  DtExamplesSidenavModule,
-  DtExamplesStepperModule,
-  DtExamplesSwitchModule,
-  DtExamplesTableModule,
-  DtExamplesTabsModule,
-  DtExamplesTagModule,
-  DtExamplesTileModule,
-  DtExamplesTimelineChartModule,
-  DtToastExamplesModule,
-  DtToggleButtonGroupExamplesModule,
-  DtExamplesTopBarNavigationModule,
-  DtExamplesTreeTableModule
-} from '@dynatrace/barista-components/examples';
-
-/**
- * NgModule that includes all example components
- */
-@NgModule({
-  imports: [
-    DtAlertExamplesModule,
-  DtAutocompleteExamplesModule,
-  DtBarIndicatorExamplesModule,
-  DtBreadcrumbsExamplesModule,
-  DtButtonExamplesModule,
-  DtButtonGroupExamplesModule,
-  DtCardExamplesModule,
-  DtChartExamplesModule,
-  DtCheckboxExamplesModule,
-  DtConfirmationDialogExamplesModule,
-  DtConsumptionExamplesModule,
-  DtContainerBreakpointObserverExamplesModule,
-  DtContextDialogExamplesModule,
-  DtCopyToClipboardExamplesModule,
-  DtCtaCardExamplesModule,
-  DtDrawerExamplesModule,
-  DtEmptyStateExamplesModule,
-  DtEventChartExamplesModule,
-  DtExpandablePanelExamplesModule,
-  DtExpandableSectionExamplesModule,
-  DtExpandableTextExamplesModule,
-  DtFilterFieldExamplesModule,
-  DtFormFieldExamplesModule,
-  DtFormattersExamplesModule,
-  DtHighlightExamplesModule,
-  DtIconExamplesModule,
-  DtInfoGroupExamplesModule,
-  DtInlineEditorExamplesModule,
-  DtInputExamplesModule,
-  DtKeyValueListExamplesModule,
-  DtLegendExamplesModule,
-  DtLinkExamplesModule,
-  DtLoadingDistractorExamplesModule,
-  DtMenuExamplesModule,
-  DtMicroChartExamplesModule,
-  DtOverlayExamplesModule,
-  DtPaginationExamplesModule,
-  DtProgressBarExamplesModule,
-  DtProgressCircleExamplesModule,
-  DtRadioExamplesModule,
-  DtExamplesSecondaryNAvModule,
-  DtExamplesSelectModule,
-  DtExamplesShowMoreModule,
-  DtExamplesSidenavModule,
-  DtExamplesStepperModule,
-  DtExamplesSwitchModule,
-  DtExamplesTableModule,
-  DtExamplesTabsModule,
-  DtExamplesTagModule,
-  DtExamplesTileModule,
-  DtExamplesTimelineChartModule,
-  DtToastExamplesModule,
-  DtToggleButtonGroupExamplesModule,
-  DtExamplesTopBarNavigationModule,
-  DtExamplesTreeTableModule
-  ]
-})
-export class DtDemosExamplesModule { }
