+++ components/package.json
@@ -24,7 +24,7 @@
   },
   "schematics": "./schematics/collection.json",
   "ng-update": {
-    "migrations": "./schematics/migration.json",
+    "migrations": "./schematics/collection.json",
     "packageGroup": [
       "@dynatrace/barista-components"
     ]
