--- components/schematics/src/migration.json
@@ -1,15 +1,0 @@
-{
-  "$schema": "../../node_modules/@angular-devkit/schematics/collection-schema.json",
-  "schematics": {
-    "migration-v5": {
-      "version": "5",
-      "description": "Updates Dynatrace Angular Components to v5",
-      "factory": "./ng-update/index#updateToV5"
-    },
-    "ng-post-update": {
-      "description": "Prints out results after ng-update.",
-      "factory": "./ng-update/index#postUpdate",
-      "private": true
-    }
-  }
-}
