--- components/schematics/src/ng-add/ast/index.ts
@@ -1,3 +1,0 @@
-export * from './read-from-tree';
-export * from './get-import-module-specifier';
-export * from './update-ngmodule-decorators';
