--- components/schematics/src/ng-add/fixtures/package-json.fixture
@@ -1,6 +1,0 @@
-{
-  "name": "angular-project",
-  "dependencies": {
-    "@dynatrace/angular-components":"4.14.0"
-  }
-}
