+++ libs/shared/utils-ast/src/lib/index.ts
@@ -14,16 +14,5 @@
  * limitations under the License.
  */
 
-export interface Schema {
-  /** Whether to install animationsmodule */
-  animations: boolean;
-
-  /** Which styles should be installed */
-  stylesPack: boolean;
-
-  /** Whether to install dynatrace iconpack */
-  iconpack: boolean;
-
-  /** Whether to setup angularJson */
-  angularJsonSetup: boolean;
-}
+export * from './get-import-module-specifier';
+export * from './update-ngmodule-decorators';
