+++ libs/shared/utils-ast/tsconfig.json
@@ -3,5 +3,8 @@
   "compilerOptions": {
     "types": ["node", "jest"]
   },
-  "include": ["**/*.ts"]
+  "include": [
+    "**/*.ts",
+    "../../../components/schematics/src/utils/read-from-tree.ts"
+  ]
 }
