+++ tools/schematics/utils/index.ts
@@ -14,16 +14,6 @@
  * limitations under the License.
  */
 
-export interface Schema {
-  /** Whether to install animationsmodule */
-  animations: boolean;
-
-  /** Which styles should be installed */
-  stylesPack: boolean;
-
-  /** Whether to install dynatrace iconpack */
-  iconpack: boolean;
-
-  /** Whether to setup angularJson */
-  angularJsonSetup: boolean;
-}
+export * from './ast-utils';
+export * from './change';
+export * from './nav-items';
